requirements.txt
pyrogram>=2.0
tgcrypto
python-dotenv
openai>=1.0.0
streamlit
openai-whisper
requests